% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_GRanges}
\alias{example_GRanges}
\title{sample GRanges coordinated with example_PlinkArray}
\format{
GRanges
}
\usage{
data("example_GRanges", package="PlinkArray")
}
\description{
sample GRanges coordinated with example_PlinkArray
}
\keyword{datasets}
