% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedPlink.R
\name{RangedPlink}
\alias{RangedPlink}
\title{constructor for RangedPlink}
\usage{
RangedPlink(delmat, ranges)
}
\arguments{
\item{delmat}{DelayedMatrix as returned by PlinkArray}

\item{ranges}{GenomicRanges checked for conformity with array}
}
\description{
constructor for RangedPlink
}
\examples{
ex = example_PlinkArray()
data("example_GRanges", package="PlinkArray")
lit = RangedPlink(ex, example_GRanges)
lit
myGR = GenomicRanges::GRanges("chr18:1-500000")
IRanges::subsetByOverlaps(lit, myGR)
}
