% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangedPlink.R
\name{subsetByOverlaps,RangedPlink,GRanges-method}
\alias{subsetByOverlaps,RangedPlink,GRanges-method}
\title{implement a form of subsetting}
\usage{
\S4method{subsetByOverlaps}{RangedPlink,GRanges}(
  x,
  ranges,
  maxgap = -1L,
  minoverlap = 0L,
  type = c("any", "start", "end", "within", "equal"),
  invert = FALSE,
  ...
)
}
\description{
implement a form of subsetting
}
